pyproject.toml
[tool.poetry]
name = "brhni-whereami"
version = "0.1.0"
description = "AFNI whereami analysis scripts"
authors = ["Bertrand Reyna-Brainerd <bertbrainerd@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
pandas = "^2.2.3"
pyyaml = "^6.0.2"

[tool.poetry.scripts]
generate-config = "brhni_whereami.generate_config:main"
whereami-analysis = "brhni_whereami.whereami_analysis:main"
analyze = "brhni_whereami.main:main" 

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
